app/requirements.txt
fastapi
uvicorn
aiogram
httpx
python-dotenv
psycopg2-binary
sqlalchemy
asyncpg
python-i18n
